backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
pydantic==2.5.2
requests==2.31.0
Pillow==10.1.0
numpy==1.24.3
tensorflow-hub==0.15.0
tensorflow==2.15.0
aiohttp==3.9.1
beautifulsoup4==4.12.2
google-generativeai==0.3.2
python-multipart==0.0.6
